requirements.txt
requests
google-generativeai
python-dotenv
tqdm
discord.py
